+++ core/class-admin-page.php
@@ -22,7 +22,6 @@ if( ! class_exists( 'Octagon_Core_Admin_Page' ) ) {
 		public $active_plugins = array();
 
 		public function __construct() {
-			add_filter( 'plugin_row_meta', array( $this, 'plugin_row_meta' ), 10, 2 );
 			add_action( 'admin_menu', array( $this, 'admin_menu' ), 99 );
 			
 		}
@@ -199,16 +198,6 @@ if( ! class_exists( 'Octagon_Core_Admin_Page' ) ) {
 
 		}
 
-		public static function plugin_row_meta( $links, $file ) {
-
-			if( OCTAGON_KC_ELEMENTS_BASENAME === $file ) {
-				$links['video-tut'] = '<a href="https://www.youtube.com/watch?v=Ot9jw7YGVEU&list=PLzuxmuoA9tTMp1iLNJHZeXjA6Rc0fqABi">' . esc_html__( 'Video Tutorial', 'octagon-kc-elements' ) . '</a>';
-				$links['support']   = '<a href="mailto:octagonwebstudio@gmail.com">' . esc_html__( 'Premium Support', 'octagon-kc-elements' ) . '</a>';
-			}
-
-			return $links;
-		}
-
 		/**
 		 * Admin Menu
 		 * 
@@ -223,6 +212,7 @@ if( ! class_exists( 'Octagon_Core_Admin_Page' ) ) {
 			
 			add_submenu_page( 'octagon-intro', esc_html_x( 'Welcome', 'admin-menu', 'octagon-kc-elements' ), esc_html_x( 'Welcome', 'admin-menu', 'octagon-kc-elements' ), 'administrator', 'octagon-intro', array( $this, 'welcome' ) );
 			add_submenu_page( 'octagon-intro', esc_html_x( 'Status', 'admin-menu', 'octagon-kc-elements' ), esc_html_x( 'Status', 'admin-menu', 'octagon-kc-elements' ), 'administrator', 'octagon-status', array( $this, 'status' ) );
+			add_submenu_page( 'octagon-intro', esc_html_x( 'Sidebar', 'admin-menu', 'octagon-kc-elements' ), esc_html_x( 'Sidebar', 'admin-menu', 'octagon-kc-elements' ), 'administrator', 'octagon-sidebar', array( $this, 'sidebar' ) );
 		}
 
 		/**
@@ -246,6 +236,17 @@ if( ! class_exists( 'Octagon_Core_Admin_Page' ) ) {
 		public function status() {
 
 			include_once OCTAGON_CORE_PATH . '/views/html-system-status.php';
+
+		}
+
+		/**
+		 * Custom Sidebar
+		 * 
+		 * @since  1.0
+		 */
+		public function sidebar() {
+
+			include_once OCTAGON_CORE_PATH . '/views/html-sidebar.php';
 
 		}